backend/requirements.txt
fastapi
uvicorn
faiss-cpu
pandas
numpy
sentence-transformers
httpx  # for calling Gemini
python-dotenv  # optional: for storing API key securely
